requirements.txt
codecov==2.1.12
coverage==6.4.4
mypy==0.971
pylint==2.15.0
pytest-aiohttp==0.3.0
pytest-cov==3.0.0
pytest-freezegun==0.4.2
pytest-socket==0.5.1
pytest-test-groups==1.0.3
pytest-sugar==0.9.5
pytest-timeout==2.1.0
pytest-xdist==2.5.0
pytest==7.1.2